+++ src/api/objects/account.ts
@@ -114,7 +114,7 @@ export default class Account implements AccountObj {
         return replaceEmojiShortcodesWithImageTags(this.displayName, this.emojis || [], fontSize);
     }
 
-    // Get the accounts instance info from the MastoApi instance
+    // Get the account's instance info from the public API (note some servers don't provide this)
     async homeInstanceInfo(): Promise<InstanceResponse> {
         const server = new MastodonServer(this.homeserver());  // Ensure the server is loaded
         return await server.fetchServerInfo();